Non-interactive run improvements
I've Resctructured the viewer code. VisIt/LibSim requires the viewer to
drive the engine. However, running non-interactively we'd like do away
with this and do everything in a single enigine process. This commit
essentially does that. The viewer is still required however now all of
our code is on the engine side. I've added two new plot examples and
cleaned up the documentation.